Allows use of longitudinal-only optimal controlled vehicle
Reorganizes some variables, notably the time tracker of when the optimal control solver is called, to allow the inclusion of longitudinal-only vehicles to the optimal controller. By extension, it also allows the inclusion of multiple vehicles with independent optimal controllers.
While at it, this commit also reorganizes the vehicle-ocp interface to prevent some code repetition.
Last, it adds the code for a platoon lane strategy